Alarm Interaction
Created light scene so all lights can be instanced. Signals are emitted to only manually connected lights, so begun creating a global messenger bus for signals, to have a signal messenger for multiple receivers (in this case lights, but can be used for the alarm generally). Global messenger can take the timer timeout and alarm snooze signals and handle those as well.